plant-disease-system/backend/requirements.txt
flask
flask-cors
tensorflow
keras
numpy
pillow
google-genai
python-dotenv
